Change HTTP timeout to 5 seconds (3 leads to timeouts)
If the iconserver hasn't been accessed in a while, Apache will close the processes.
If a process needs to be started, connections will often timeout after 3 seconds.

Similarly, for uses with Little Snitch installed, 3 seconds isn't a whole lot of time to
read the dialog and decide if you want to allow the connection.